requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
python-dotenv==1.0.0
psutil==5.9.6
aiohttp==3.9.1
aiofiles==23.2.1
requests==2.31.0
pillow==10.1.0
httpx==0.25.2
uvloop==0.19.0